métodos DAO añadidos

+++ src/main/java/org/BBDD/DAO.java
@@ -3,12 +3,12 @@ package org.BBDD;
 import java.util.List;
 
 public interface DAO <T>{
-    T get(long id);
+    T get(int id);
     List<T> getAll();
     void save(T t);
     void update(T t);
     void delete(T t);
-    boolean deleteById(long id);
+    boolean deleteById(int id);
     List<Integer> getAllIds();
     void updateLOB(T book, String f);
     void updateLOBById(long id, String f);
